enhance virtualenvsetup to not require python-virtualenv installed
depending on python-virtualenv on the host is hard to make work
as the virtualenv version might be different with maybe bad support for https, etc.

This makes the setup dependent on wget and unzip though
