+++ tox.ini
@@ -5,9 +5,10 @@ basepython = python2.7
 
 [testenv:config]
 deps=
-    buildbot<9
+    buildbot<0.9.0
 setenv=
     CHECK_CONFIG=true
+    PYTHONPATH=.
 changedir=
     {envtmpdir}
 whitelist_externals=ln
@@ -15,6 +16,7 @@ commands=
     ln -s {toxinidir} metabbotcfg
     ln -s {toxinidir}/master.cfg master.cfg
     buildbot checkconfig
+    trial metabbotcfg
 
 [testenv:lint]
 deps=
